Tsavorite Native: address GPT-5.5 PR review findings
Fixes two correctness issues caught by an automated code review of the
optimize-device PR.

### io_uring SQE leak on submit failure

In UringFile::ScheduleOperation, io_uring_get_sqe() advances the
user-side sqe_tail before io_uring_submit() is called. If submit fails
after retries (-EAGAIN/-EBUSY exhausted, or any other negative), the
old code released the lock and returned IOError without doing anything
about the still-pending SQE. user_data on that SQE pointed at the
io_context unique_ptr that was about to be freed by the guards
unwinding, so the next successful submit on the same ring would consume
the stale SQE and the QueueRunFor drain loop would dispatch a callback
against freed memory — a clear use-after-free.

Fix: before releasing sq_lock on the failure path, rewrite the still-
pending SQE in place as io_uring_prep_nop with user_data = nullptr.
The drain loop already skips nullptr user_data (it's the wake-up
sentinel used by UringIoHandler::Wake), so when a later submit flushes
this nop the CQE is drained harmlessly. Safe to mutate the SQE in place
because we still hold sq_lock and no kernel/concurrent submitter has
observed it yet.

### NativeDevice sector_size always returned 512

FileSystemSegmentedFile::alignment() returned a hard-coded 512.
NativeDeviceImpl::sector_size() delegated to it, so the C# wrapper's
sector-size cross-check in EnsureNativeDeviceCreated would:

  - falsely throw on 4K-native disks where ProbeAlignment returns 4096
    (managed 4096 vs native 512 → 'sector-size mismatch' → device
    unusable), or
  - on 4K-native disks where the managed probe fell back to 512 (e.g.
    older kernel without STATX_DIOALIGN), let the device initialize
    with SectorSize=512 and then have the kernel reject the 512-aligned
    O_DIRECT buffers with EINVAL.

Fix: factor the STATX_DIOALIGN probe from NativeDevice_ProbeAlignment
into a shared inline helper (native_device::ProbeDioAlignment in
native_device.h) and call it once from the NativeDeviceImpl ctor,
caching the result as the immutable member device_alignment_.
sector_size() now returns the cached value; NativeDevice_ProbeAlignment
delegates to the same helper. Both sides of the ABI cross-check go
through identical probe logic, so the check is now a meaningful ABI /
runtime-drift detector instead of a 4K-disk footgun.

### Stale IDevice.Initialize XML

The omitSegmentIdFromFilename param said it was 'only supported by
managed devices — NativeStorageDevice rejects this flag'. Native
devices have honored the flag since 6584cf7. Updated the doc.

### Alpine install hint

The 'IoBackend.Uring with libaio fallback' error message suggested
'sudo apk add liburing' on Alpine, but README.md notes that the
prebuilt won't load on Alpine (musl) at all. Replaced the apk
suggestion with the actual Alpine support story (use a glibc image or
fall back to a managed device).

Verification: 62/62 Tsavorite.test.hlog IDevice + NativeStorageDevice
tests pass on Linux. Both .so binaries rebuilt (uring-enabled and
libaio-only fallback) with correct ldd output. Device.benchmark NVMe
saturation throughput unchanged within noise (libaio 738K IOPS, uring
349K IOPS on Dell P5600).
